+++ packages/web/src/pages/notification-users-page/NotificationUsersPage.tsx
@@ -11,7 +11,7 @@ import { Dispatch } from 'redux'
 
 import MobilePageContainer from 'components/mobile-page-container/MobilePageContainer'
 import NavContext, { LeftPreset } from 'components/nav/mobile/NavContext'
-import UserList from 'components/user-list/UserList'
+import { UserList } from 'components/user-list/UserList'
 import { AppState } from 'store/types'
 const { getPageTitle, getUserList } = notificationsUserListSelectors
 const { loadMore } = userListActions
